Fix #23090: Added style option to use expression font for dynamics
This commit introduces a new class DynamicFragment which inherits the TextFragment class. This is used to handle the changes for Dynamics.

+++ src/engraving/dom/harppedaldiagram.cpp
@@ -140,7 +140,7 @@ void HarpPedalDiagram::setIsDiagram(bool diagram)
 
 void HarpPedalDiagram::setPedal(HarpStringType harpString, PedalPosition pedal)
 {
-    m_pedalState.at(harpString) = pedal;
+    m_pedalState.at(static_cast<size_t>(harpString)) = pedal;
     setPlayableTpcs();
 }
